fix: restructure and remove barrel file
