+++ packages/design-system/src/components/N8nKeyboardShortcut/N8nKeyboardShortcut.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useDeviceSupport } from '@n8n/frontend-composables';
+import { useDeviceSupport } from '@n8n/frontend-composables/useDeviceSupport';
 import { computed } from 'vue';
 
 import type { KeyboardShortcut } from '../../types/keyboardshortcut';
