+++ packages/editor-ui/src/components/canvas/Canvas.test.ts
@@ -6,7 +6,7 @@ import { createPinia, setActivePinia } from 'pinia';
 import type { CanvasConnection, CanvasNode } from '@/types';
 import { createCanvasConnection, createCanvasNodeElement } from '@/__tests__/data';
 import { NodeConnectionType } from 'n8n-workflow';
-import type { useDeviceSupport } from '@n8n/frontend-composables';
+import type { useDeviceSupport } from '@n8n/frontend-composables/useDeviceSupport';
 
 const matchMedia = global.window.matchMedia;
 // @ts-expect-error Initialize window object
