+++ packages/editor-ui/tsconfig.json
@@ -12,7 +12,7 @@
 		"paths": {
 			"@/*": ["./src/*"],
 			"n8n-design-system*": ["../design-system/src*"],
-			"@n8n/frontend-composables*": ["../frontend/composables/src*"],
+			"@n8n/frontend-composables*": ["../frontend/@n8n/composables/src*"],
 			"@n8n/chat/*": ["../@n8n/chat/src/*"],
 			"@n8n/api-types*": ["../@n8n/api-types/src*"]
 		},
