+++ packages/frontend/@n8n/composables/.eslintrc.cjs
