+++ packages/frontend/@n8n/composables/biome.jsonc
@@ -1,0 +1,4 @@
+{
+	"$schema": "../../../../node_modules/@biomejs/biome/configuration_schema.json",
+	"extends": ["../../../../biome.jsonc"]
+}
