+++ packages/frontend/@n8n/composables/package.json
@@ -1,0 +1,46 @@
+{
+	"name": "@n8n/frontend-composables",
+	"type": "module",
+	"files": [
+		"dist"
+	],
+	"exports": {
+		"./*": {
+			"types": "./dist/*.d.ts",
+			"import": "./dist/*.js",
+			"require": "./dist/*.cjs"
+		}
+	},
+	"scripts": {
+		"dev": "vite",
+		"build": "pnpm run typecheck && tsup",
+		"preview": "vite preview",
+		"typecheck": "vue-tsc --noEmit",
+		"test": "vitest run",
+		"test:dev": "vitest",
+		"lint": "eslint src --ext .js,.ts,.vue --quiet",
+		"lintfix": "eslint src --ext .js,.ts,.vue --fix",
+		"format": "biome format --write . && prettier --write . --ignore-path ../../../.prettierignore",
+		"format:check": "biome ci . && prettier --check . --ignore-path ../../../.prettierignore"
+	},
+	"dependencies": {
+		"vue": "catalog:frontend"
+	},
+	"devDependencies": {
+		"@n8n/frontend-eslint-config": "workspace:*",
+		"@n8n/frontend-typescript-config": "workspace:*",
+		"@n8n/frontend-vitest-config": "workspace:*",
+		"@testing-library/jest-dom": "catalog:frontend",
+		"@testing-library/user-event": "catalog:frontend",
+		"@testing-library/vue": "catalog:frontend",
+		"@vitejs/plugin-vue": "catalog:frontend",
+		"@vue/tsconfig": "catalog:frontend",
+		"tsup": "catalog:frontend",
+		"typescript": "catalog:frontend",
+		"vite": "catalog:frontend",
+		"vite-plugin-dts": "catalog:frontend",
+		"vitest": "catalog:frontend",
+		"vue-tsc": "catalog:frontend"
+	},
+	"license": "See LICENSE.md file in the root of the repository"
+}
