+++ packages/frontend/@n8n/composables/src/useDeviceSupport.ts
