+++ packages/frontend/@n8n/composables/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-	entry: ['src/index.ts'],
+	entry: ['src/useDeviceSupport.ts'],
 	format: ['cjs', 'esm'],
 	clean: true,
 	dts: true,
