--- packages/frontend/composables/biome.jsonc
@@ -1,4 +1,0 @@
-{
-	"$schema": "../../../node_modules/@biomejs/biome/configuration_schema.json",
-	"extends": ["../../../biome.jsonc"]
-}
